requirements.txt
dbus-python
pygobject